refactor: utils and createDrive helper
