+++ src/modules/filemanager/pages/main/MainPage.tsx
@@ -7,7 +7,7 @@ import { FileBrowser } from '../../components/FileBrowser/FileBrowser'
 import { FMFileViewProvider } from '../../providers/FMFileViewContext'
 import { FMInitialModal } from '../../components/FMInitialModal/FMInitialModal'
 import { FMProvider } from '../../providers/FMContext'
-import { getUsableStamps } from '../../utils/utils'
+import { getUsableStamps } from '../../utils/bee'
 import { Context as SettingsContext, Provider as SettingsProvider } from '../../../../providers/Settings'
 import { PostageBatch } from '@ethersphere/bee-js'
 import { FMSearchProvider } from '../../providers/FMSearchContext'
