--- src/modules/filemanager/utils/fm/index.ts
@@ -1,3 +1,0 @@
-export * from './indexes'
-export * from './publishers'
-export * from './ui'
